Renamed DTO classes to Response suffix for clarity and updated references across the network and data layers.
